feat(ui): replace versions bar with floating links
- add floating links component and update resume layout
- remove versions bar component
- refresh page styles for the new layout
- update README accordingly
