+++ assets/page.css
@@ -175,48 +175,41 @@ svg {
   text-align: center;
 }
 
-.versions-bar {
+.floating-links {
+  position: fixed;
+  right: 1.4em;
+  bottom: 1.4em;
+  display: flex;
+  flex-direction: column;
+  gap: 0.6em;
+  z-index: 110;
+}
+
+.floating-links a {
+  display: grid;
+  place-items: center;
+  width: 2.6em;
+  height: 2.6em;
+  border-radius: 50%;
   background: var(--color-dimmed);
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.75em 1.5em;
-  align-items: center;
-  justify-content: center;
-  grid-column: full;
-  padding: 0.6em 1em;
-  font-size: 0.9em;
-}
-
-.versions-bar__group {
-  display: flex;
-  align-items: center;
-  gap: 0.6em;
-}
-
-.versions-bar__label {
-  color: var(--color-secondary);
-  font-size: 0.75em;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-}
-
-.versions-bar__list {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  display: flex;
-  gap: 0.6em;
-  flex-wrap: wrap;
-}
-
-.versions-bar__list a {
   color: var(--color-accent);
   text-decoration: none;
-  font-weight: 600;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+  transition:
+    transform 0.2s ease,
+    box-shadow 0.2s ease;
 }
 
-.versions-bar__list a:hover {
-  text-decoration: underline;
+.floating-links a:hover,
+.floating-links a:focus {
+  transform: translateY(-2px);
+  box-shadow: 0 6px 18px rgba(0, 0, 0, 0.2);
+}
+
+@media print {
+  .floating-links {
+    display: none;
+  }
 }
 
 .masthead > *,
